rl_system/scripts/train_hrl_pretrain.py
#!/usr/bin/env python3
"""
HRL Specialist Pre-Training Script (Phase 1: Stub)

Entrypoint for pre-training individual specialists (SEARCH, TRACK, TERMINAL).

Phase 1: Validates setup and prints readiness message.
Phase 2: Will implement actual PPO training for each specialist.
"""

import sys
import argparse
import yaml
from pathlib import Path

# Add parent directory to path
sys.path.insert(0, str(Path(__file__).parent.parent))


def parse_args():
    """Parse command line arguments."""
    parser = argparse.ArgumentParser(
        description="Pre-train HRL specialist policies",
        formatter_class=argparse.RawDescriptionHelpFormatter,
        epilog="""
Examples:
  # Pre-train search specialist
  python scripts/train_hrl_pretrain.py --agent search --config configs/hrl/search_specialist.yaml

  # Pre-train track specialist
  python scripts/train_hrl_pretrain.py --agent track --config configs/hrl/track_specialist.yaml

  # Pre-train terminal specialist
  python scripts/train_hrl_pretrain.py --agent terminal --config configs/hrl/terminal_specialist.yaml
        """
    )

    parser.add_argument(
        "--agent",
        type=str,
        required=True,
        choices=["search", "track", "terminal"],
        help="Which specialist to pre-train"
    )

    parser.add_argument(
        "--config",
        type=str,
        required=True,
        help="Path to specialist config YAML"
    )

    parser.add_argument(
        "--steps",
        type=int,
        default=None,
        help="Override total training steps"
    )

    parser.add_argument(
        "--checkpoint-dir",
        type=str,
        default=None,
        help="Override checkpoint directory"
    )

    return parser.parse_args()


def validate_config(config_path: Path) -> dict:
    """Load and validate config."""
    if not config_path.exists():
        raise FileNotFoundError(f"Config not found: {config_path}")

    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)

    return config


def pretrain_specialist_stub(agent: str, config: dict):
    """
    Stub implementation: prints readiness message.

    Phase 2: This will:
    1. Create specialized environment wrapper for the agent's phase
    2. Initialize PPO with specialist config
    3. Train for specified steps
    4. Save specialist checkpoint
    """
    print("\n" + "=" * 60)
    print(f"HRL Specialist Pre-Training: {agent.upper()}")
    print("=" * 60)

    print(f"\n✅ Configuration loaded successfully")
    print(f"   Agent: {agent}")
    print(f"   Training steps: {config.get('training', {}).get('total_timesteps', 'N/A')}")
    print(f"   Network arch: {config.get('training', {}).get('net_arch', 'N/A')}")
    print(f"   Learning rate: {config.get('training', {}).get('learning_rate', 'N/A')}")

    print(f"\n🚧 Phase 1 Stub: No training logic implemented yet")
    print(f"   TODO(Phase 2): Implement specialist-specific environment wrapper")
    print(f"   TODO(Phase 2): Initialize PPO with specialist config")
    print(f"   TODO(Phase 2): Train for specified steps")
    print(f"   TODO(Phase 2): Save checkpoint to checkpoints/hrl/specialists/{agent}/")

    print(f"\n✓ Setup validated - ready for Phase 2 implementation")


def main():
    """Main entrypoint."""
    args = parse_args()

    print("\n🚀 HRL Specialist Pre-Training (Phase 1: Stub)")

    # Validate config
    config_path = Path(args.config)
    try:
        config = validate_config(config_path)
    except Exception as e:
        print(f"\n❌ Error loading config: {e}")
        return 1

    # Run stub training
    try:
        pretrain_specialist_stub(args.agent, config)
        print("\n✅ Pre-training stub completed successfully\n")
        return 0
    except Exception as e:
        print(f"\n❌ Error during pre-training: {e}")
        import traceback
        traceback.print_exc()
        return 1


if __name__ == "__main__":
    sys.exit(main())
